DSND-2169: Refactor annotations mapper
