+++ src/main/java/uk/gov/companieshouse/filinghistory/api/mapper/response/AnnotationsResponseMapper.java
@@ -1,0 +1,26 @@
+package uk.gov.companieshouse.filinghistory.api.mapper.response;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+import org.springframework.stereotype.Component;
+import uk.gov.companieshouse.api.filinghistory.FilingHistoryItemDataAnnotations;
+import uk.gov.companieshouse.filinghistory.api.model.FilingHistoryAnnotation;
+
+@Component
+public class AnnotationsResponseMapper {
+
+    public List<FilingHistoryItemDataAnnotations> map(List<FilingHistoryAnnotation> documentAnnotations) {
+
+        Optional<List<FilingHistoryAnnotation>> optionalDocumentAnnotations = Optional.ofNullable(documentAnnotations);
+
+        return optionalDocumentAnnotations.map(inputAnnotations -> {
+            List<FilingHistoryItemDataAnnotations> outputAnnotations = new ArrayList<>();
+            inputAnnotations.forEach(annotation ->
+                    outputAnnotations.add(
+                            new FilingHistoryItemDataAnnotations()
+                                    .annotation(annotation.getAnnotation())));
+            return outputAnnotations;
+        }).orElse(null);
+    }
+}
