+++ src/main/java/uk/gov/companieshouse/filinghistory/api/service/Processor.java
@@ -1,6 +1,5 @@
 package uk.gov.companieshouse.filinghistory.api.service;
 
-import javax.naming.ServiceUnavailableException;
 import uk.gov.companieshouse.api.filinghistory.ExternalData;
 import uk.gov.companieshouse.api.filinghistory.InternalFilingHistoryApi;
